Mewen Correction Blanche

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: apolleux <apolleux@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/13 18:23:41 by apolleux          #+#    #+#             */
-/*   Updated: 2025/10/28 14:28:33 by apolleux         ###   ########.fr       */
+/*   Updated: 2025/10/29 17:48:38 by apolleux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
